chore(action-parser): solve circular dependency by rename file
